% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diag_plots.R
\name{Diag_plots}
\alias{Diag_plots}
\title{Diagnostic Plots for Model Fits}
\usage{
Diag_plots(fits, stk, idx, outdir = "plots_diagnostics", prefix = "fit")
}
\arguments{
\item{fits}{A list of model fits (e.g., from loop_scas or custom fmodels).}

\item{stk}{FLStock object (used in residuals and fit plots).}

\item{idx}{FLIndex object (used in residuals and fit plots).}

\item{outdir}{Folder where plots will be saved (default: "plots/diagnostics").}

\item{prefix}{Optional filename prefix.}
}
\description{
Diagnostic Plots for Model Fits
}
