% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_loop_scas.R
\name{dynamic_loop_scas}
\alias{dynamic_loop_scas}
\title{Generate fits from dynamic fmodels based on a k-range}
\usage{
dynamic_loop_scas(stk, idx, k_range = 1:20, qmodel, srmodel)
}
\arguments{
\item{stk}{FLStock}

\item{idx}{FLIndex}

\item{k_range}{A vector of k values (default = 1:20)}

\item{qmodel}{qmodel}

\item{srmodel}{srmodel}
}
\value{
List of fits
}
\description{
Generate fits from dynamic fmodels based on a k-range
}
