% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_formula.R
\name{f}
\alias{f}
\title{Construct GAM Formula Based on String Code}
\usage{
f(model_string, ks)
}
\arguments{
\item{model_string}{A model string (e.g., "sy", "sa", or "sy+sa").}

\item{ks}{List of node values (must match the number of model terms).}
}
\value{
A formula object.
}
\description{
Construct GAM Formula Based on String Code
}
