% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_scas.R
\name{loop_scas}
\alias{loop_scas}
\title{Loop Over SCA Fits for Different Model Complexities}
\usage{
loop_scas(stk, idx, expression = "sy", qmodel, srmodel)
}
\arguments{
\item{stk}{FLStock object.}

\item{idx}{FLIndex object.}

\item{expression}{Model expression string ("sy", "sa", or "sy+sa").}

\item{qmodel}{Q model object.}

\item{srmodel}{SR model object.}
}
\value{
List of fits from scas().
}
\description{
Loop Over SCA Fits for Different Model Complexities
}
