% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_summary.R
\name{model_summary}
\alias{model_summary}
\title{Generate Model Fit Summary (AIC, BIC, GCV)}
\usage{
model_summary(fits, ks = NULL)
}
\arguments{
\item{fits}{A list of fits from loop_scas().}

\item{ks}{Optional vector of model complexity values.}
}
\value{
A data.frame with k, AIC, BIC, GCV.
}
\description{
Generate Model Fit Summary (AIC, BIC, GCV)
}
