% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohn.R
\name{mohn}
\alias{mohn}
\title{Compute Mohn's Rho}
\usage{
mohn(retro.lst, qoi = c("fbar", "ssb", "rec"))
}
\arguments{
\item{retro.lst}{A list of FLStock objects (output from \code{retro()}).}

\item{qoi}{Quantities of interest: one or more of "fbar", "ssb", "rec".}
}
\value{
Named numeric vector of Mohn's rho values.
}
\description{
Compute Mohn's Rho
}
