% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro.R
\name{retro}
\alias{retro}
\title{Perform Retrospective Analysis}
\usage{
retro(stk, idxs, retro = 4, kfrac = 0.4, k, ...)
}
\arguments{
\item{stk}{FLStock object.}

\item{idxs}{FLIndices object.}

\item{retro}{Number of retrospective years.}

\item{kfrac}{Fraction of nodes to remove per step (default = 0.4).}

\item{k}{Number of nodes.}

\item{...}{Additional arguments passed to \code{sca()}.}
}
\value{
FLStocks object with retrospective fits.
}
\description{
Perform Retrospective Analysis
}
